// src/lib/scraper/types.ts
export interface Business {
  id: string;
  maps_address: string;
  name: string;
  phone: string;
  provider: string;
  address: string;
  type_of_business: string;
  town: string;
  notes: string;
}
// ...
export interface ExportExcelRequest {
  businesses: Business[];
  filename: string;
  addHyperlinks?: boolean;
}